requirements.txt
Pygments
polib
requests
cherrypy
pyyaml
-e git://github.com/sloonz/pil-py3k.git#egg=pil
misaka